Split out views/operations.py
