+++ corehq/apps/hqadmin/views/views.py
@@ -195,28 +195,6 @@ class SuperuserManagement(BaseAdminSectionView):
         return self.get(request, *args, **kwargs)
 
 
-@require_superuser_or_contractor
-def mass_email(request):
-    if request.method == "POST":
-        form = EmailForm(request.POST)
-        if form.is_valid():
-            subject = form.cleaned_data['email_subject']
-            html = form.cleaned_data['email_body_html']
-            text = form.cleaned_data['email_body_text']
-            real_email = form.cleaned_data['real_email']
-            send_mass_emails.delay(request.couch_user.username, real_email, subject, html, text)
-            messages.success(request, 'Task started. You will receive an email summarizing the results.')
-        else:
-            messages.error(request, 'Something went wrong, see below.')
-    else:
-        form = EmailForm()
-
-    context = get_hqadmin_base_context(request)
-    context['hide_filters'] = True
-    context['form'] = form
-    return render(request, "hqadmin/mass_email.html", context)
-
-
 class AdminRestoreView(TemplateView):
     template_name = 'hqadmin/admin_restore.html'
 
@@ -641,85 +619,6 @@ class DisableTwoFactorView(FormView):
         return redirect('{}?q={}'.format(reverse('web_user_lookup'), username))
 
 
-@require_superuser
-def callcenter_test(request):
-    user_id = request.GET.get("user_id")
-    date_param = request.GET.get("date")
-    enable_caching = request.GET.get('cache')
-    doc_id = request.GET.get('doc_id')
-
-    if not user_id and not doc_id:
-        return render(request, "hqadmin/callcenter_test.html", {"enable_caching": enable_caching})
-
-    error = None
-    user = None
-    user_case = None
-    domain = None
-    if user_id:
-        try:
-            user = CommCareUser.get(user_id)
-            domain = user.project
-        except ResourceNotFound:
-            error = "User Not Found"
-    elif doc_id:
-        try:
-            doc = CommCareUser.get_db().get(doc_id)
-            domain = Domain.get_by_name(doc['domain'])
-            doc_type = doc.get('doc_type', None)
-            if doc_type == 'CommCareUser':
-                case_type = domain.call_center_config.case_type
-                user_case = CaseAccessors(doc['domain']).get_case_by_domain_hq_user_id(doc['_id'], case_type)
-            elif doc_type == 'CommCareCase':
-                if doc.get('hq_user_id'):
-                    user_case = CommCareCase.wrap(doc)
-                else:
-                    error = 'Case ID does does not refer to a Call Center Case'
-        except ResourceNotFound:
-            error = "User Not Found"
-
-    try:
-        query_date = dateutil.parser.parse(date_param)
-    except ValueError:
-        error = "Unable to parse date, using today"
-        query_date = date.today()
-
-    def view_data(case_id, indicators):
-        new_dict = OrderedDict()
-        key_list = sorted(indicators)
-        for key in key_list:
-            new_dict[key] = indicators[key]
-        return {
-            'indicators': new_dict,
-            'case': CommCareCase.get(case_id),
-        }
-
-    if user or user_case:
-        custom_cache = None if enable_caching else cache.caches['dummy']
-        override_case = CallCenterCase.from_case(user_case)
-        cci = CallCenterIndicators(
-            domain.name,
-            domain.default_timezone,
-            domain.call_center_config.case_type,
-            user,
-            custom_cache=custom_cache,
-            override_date=query_date,
-            override_cases=[override_case] if override_case else None
-        )
-        data = {case_id: view_data(case_id, values) for case_id, values in cci.get_data().items()}
-    else:
-        data = {}
-
-    context = {
-        "error": error,
-        "mobile_user": user,
-        "date": json_format_date(query_date),
-        "enable_caching": enable_caching,
-        "data": data,
-        "doc_id": doc_id
-    }
-    return render(request, "hqadmin/callcenter_test.html", context)
-
-
 class DownloadMALTView(BaseAdminSectionView):
     urlname = 'download_malt'
     page_title = ugettext_lazy("Download MALT")
@@ -796,109 +695,12 @@ def _gir_csv_response(month, year):
     return response
 
 
-class CallcenterUCRCheck(BaseAdminSectionView):
-    urlname = 'callcenter_ucr_check'
-    page_title = ugettext_lazy("Check Callcenter UCR tables")
-    template_name = "hqadmin/call_center_ucr_check.html"
-
-    @method_decorator(require_superuser)
-    def dispatch(self, request, *args, **kwargs):
-        return super(CallcenterUCRCheck, self).dispatch(request, *args, **kwargs)
-
-    @property
-    def page_context(self):
-        from corehq.apps.callcenter.data_source import get_call_center_domains
-        from corehq.apps.callcenter.checks import get_call_center_data_source_stats
-
-        if 'domain' not in self.request.GET:
-            return {}
-
-        domain = self.request.GET.get('domain', None)
-        if domain:
-            domains = [domain]
-        else:
-            domains = [dom.name for dom in get_call_center_domains() if dom.use_fixtures]
-
-        domain_stats = get_call_center_data_source_stats(domains)
-
-        context = {
-            'data': sorted(list(domain_stats.values()), key=lambda s: s.name),
-            'domain': domain
-        }
-
-        return context
-
-
 class DimagisphereView(TemplateView):
 
     def get_context_data(self, **kwargs):
         context = super(DimagisphereView, self).get_context_data(**kwargs)
         context['tvmode'] = 'tvmode' in self.request.GET
         return context
-
-
-class ReprocessMessagingCaseUpdatesView(BaseAdminSectionView):
-    urlname = 'reprocess_messaging_case_updates'
-    page_title = ugettext_lazy("Reprocess Messaging Case Updates")
-    template_name = 'hqadmin/messaging_case_updates.html'
-
-    @method_decorator(require_superuser)
-    def dispatch(self, request, *args, **kwargs):
-        return super(ReprocessMessagingCaseUpdatesView, self).dispatch(request, *args, **kwargs)
-
-    @property
-    @memoized
-    def form(self):
-        if self.request.method == 'POST':
-            return ReprocessMessagingCaseUpdatesForm(self.request.POST)
-        return ReprocessMessagingCaseUpdatesForm()
-
-    @property
-    def page_context(self):
-        context = get_hqadmin_base_context(self.request)
-        context.update({
-            'form': self.form,
-        })
-        return context
-
-    def get_case(self, case_id):
-        try:
-            return CaseAccessorSQL.get_case(case_id)
-        except CaseNotFound:
-            pass
-
-        try:
-            return CaseAccessorCouch.get_case(case_id)
-        except ResourceNotFound:
-            pass
-
-        return None
-
-    def post(self, request, *args, **kwargs):
-        from corehq.messaging.signals import messaging_case_changed_receiver
-
-        if self.form.is_valid():
-            case_ids = self.form.cleaned_data['case_ids']
-            case_ids_not_processed = []
-            case_ids_processed = []
-            for case_id in case_ids:
-                case = self.get_case(case_id)
-                if not case or case.doc_type != 'CommCareCase':
-                    case_ids_not_processed.append(case_id)
-                else:
-                    messaging_case_changed_receiver(None, case)
-                    case_ids_processed.append(case_id)
-
-            if case_ids_processed:
-                messages.success(self.request,
-                    _("Processed the following case ids: {}").format(','.join(case_ids_processed)))
-
-            if case_ids_not_processed:
-                messages.error(self.request,
-                    _("Could not find cases belonging to these case ids: {}")
-                    .format(','.join(case_ids_not_processed)))
-
-        return self.get(request, *args, **kwargs)
 
 
 def top_five_projects_by_country(request):
